More sane types, TaskContext and misc fixes

+++ serial_comm/SerialCommTask.cpp
@@ -2,13 +2,13 @@
 
 #include "Arduino.h"
 
-SerialCommTask::SerialCommTask(const unsigned int runPeriodicity, const long baudRate) : 
+SerialCommTask::SerialCommTask(const uint16_t runPeriodicity, const unsigned long baudRate) : 
 TaskBase(runPeriodicity)
 {
   Serial.begin(baudRate);
 }
 
-void SerialCommTask::execute(const unsigned int newTime)
+void SerialCommTask::execute(const unsigned long newTime)
 {
   if (0 < Serial.availableForWrite())
   {
